+++ R/dx_heart_failure.R
@@ -8,7 +8,8 @@
 #'   \item{AgeGroup}{Age group}
 #'   \item{Sex}{sex}
 #'   \item{truth}{Heart failure (outcome)}
-#'   \item{predicted}{Predicted outcome from model}
+#'   \item{predicted}{Predicted outcome from a GLM model}
+#'   \item{predicted_rf}{Predicted outcome from a Random Forest model}
 #'   \item{AgeSex}{Age and sex group}
 #' }
 #' @source \url{https://www.kaggle.com/imnikhilanand/heart-attack-prediction}